httpfunctions/n8n-python/requirements.txt
fastapi==0.115.0
cloudbase-agent-core==0.1.13
cloudbase-agent-server==0.1.22
cloudbase-agent-n8n==0.1.2
cloudbase-agent-observability==0.1.3
ag-ui-protocol==0.1.10
python-dotenv==1.0.1
zstandard>=0.22.0
